Add x_page to example
