--- example/ios_example/AppState.swift
@@ -1,11 +1,0 @@
-import SwiftUI
-
-struct UserInfo {
-    var userID: String
-    var isAnonymous: Bool
-    var isVerified: Bool
-}
-
-class AppState: ObservableObject {
-    @Published var user: UserInfo?
-}
